Restructure services into backend and client layers
