+++ src/dc43/components/contract_validation/core.py
