+++ src/dc43/components/data_quality/governance/interface.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, Mapping, Optional, Protocol, Sequence
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.components.data_quality.engine import ValidationResult
+from dc43.components.contract_validation import ValidationResult
 
 
 @dataclass
