+++ src/dc43/components/data_quality/manager.py
@@ -4,9 +4,11 @@ from __future__ import annotations
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
+from dc43.components.contract_validation import (
+    ValidationResult,
+    evaluate_observations,
+)
 from dc43.lib.data_quality import ObservationPayload
-
-from .engine import ValidationResult, evaluate_observations
 
 
 class DataQualityManager:
