+++ src/dc43/demo_app/pipeline.py
@@ -23,7 +23,7 @@ from dc43.demo_app.server import (
 )
 from dc43.components.data_quality.governance import DQStatus
 from dc43.components.data_quality.integration import attach_failed_expectations
-from dc43.services.governance.local import build_local_governance_service
+from dc43.services.governance.client import build_local_governance_service
 from dc43.integration.spark_io import (
     ContractFirstDatasetLocator,
     ContractVersionLocator,
