+++ src/dc43/services/contracts/backend/__init__.py
@@ -1,0 +1,6 @@
+"""Backend contracts and stubs for contract management services."""
+
+from .interface import ContractServiceBackend
+from .local import LocalContractServiceBackend
+
+__all__ = ["ContractServiceBackend", "LocalContractServiceBackend"]
