+++ src/dc43/services/contracts/client/__init__.py
@@ -1,0 +1,6 @@
+"""Client interfaces and local implementations for contract services."""
+
+from .interface import ContractServiceClient
+from .local import LocalContractServiceClient
+
+__all__ = ["ContractServiceClient", "LocalContractServiceClient"]
