+++ src/dc43/services/data_quality/backend/__init__.py
@@ -1,0 +1,6 @@
+"""Backend contracts and stubs for the data-quality service."""
+
+from .interface import DataQualityServiceBackend
+from .local import LocalDataQualityServiceBackend
+
+__all__ = ["DataQualityServiceBackend", "LocalDataQualityServiceBackend"]
