+++ src/dc43/services/governance/__init__.py
@@ -1,7 +1,11 @@
 """Governance orchestration service abstractions and local stubs."""
 
-from .client import GovernanceServiceClient
-from .local import LocalGovernanceService, build_local_governance_service
+from .backend import GovernanceServiceBackend, LocalGovernanceServiceBackend
+from .client import (
+    GovernanceServiceClient,
+    LocalGovernanceServiceClient,
+    build_local_governance_service,
+)
 from .models import (
     GovernanceCredentials,
     PipelineContext,
@@ -11,7 +15,11 @@ from .models import (
     normalise_pipeline_context,
 )
 
+LocalGovernanceService = LocalGovernanceServiceBackend
+
 __all__ = [
+    "GovernanceServiceBackend",
+    "LocalGovernanceServiceBackend",
     "GovernanceServiceClient",
     "GovernanceCredentials",
     "PipelineContext",
@@ -19,6 +27,7 @@ __all__ = [
     "QualityAssessment",
     "QualityDraftContext",
     "normalise_pipeline_context",
+    "LocalGovernanceServiceClient",
     "LocalGovernanceService",
     "build_local_governance_service",
 ]
