+++ src/dc43/services/governance/backend/__init__.py
@@ -1,0 +1,6 @@
+"""Backend contracts and stubs for governance services."""
+
+from .interface import GovernanceServiceBackend
+from .local import LocalGovernanceServiceBackend
+
+__all__ = ["GovernanceServiceBackend", "LocalGovernanceServiceBackend"]
